Documentation: Make the toctree more web-friendly
The toctree is displayed in the navigation bar at the top of the page.
Add short names to the entries that replicate the names currently used
on libcamera.org.

Remove the link to the index and search pages, as the former is empty,
and the latter can be accessed directory from the navigation bar search
box.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>