Documentation: Add placeholder for Doxygen-generated API documentation
Until we have better Sphinx + Doxygen integration, replace the relative
external link to the API documentation with a placeholder directory,
which will be populated with the Doxygen-generated doc when exporting
documentation to the website. This is needed as the Sphinx documentation
is exported to the root of the website, linking to ../api-html is thus
not possible.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Acked-by: Kieran Bingham <kieran.bingham@ideasonboard.com>