libcamera: camera: Fix operator= definition
The class assignment operator is usually defined as returning a
reference to the object, to allow a = b = c statements. While the return
type makes no difference when deleting the operator in a class
definition, it's still a good practice to use the correct return type.
Fix it in the Camera and CameraManager classes.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>