test: camera: Add capture test
Correctly configure the camera using the default configuration and run a
capture session for 100 milliseconds, which is plenty of time, in tests
over 600 requests completed using the vimc pipeline.

The test passes if at least the number of buffers used in the capture
times two number of requests completes to prove we cycle through all
buffers.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>