test: camera: Add state machine test
Add a test of the different access level enforced by the state machine
inside the camera. The state machine aims to limit operations on the
camera to the cameras state.

The test exercises all states of the camera and verifies that only the
intended operations are possible at each stage.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>